WIP: split server-attendees.test.ts into themed files — NOT GREEN, cpd has 49 clones (0.3%)
Split the 3674-line monolith (grandfathered in biome.json's
noExcessiveLinesPerFile override) into 12 themed files under
test/lib/server-attendees/, each under the 1000-line Biome hard limit:
delete, delete-incomplete, checkin, add-attendee, attendee-detail,
attendee-edit, resend-notification, payment, questions, merge-panel,
merge-post, merge-conflicts (+ helpers.ts with getMergeVersion).

Pure relocation via verbatim line-range extraction; no assertions changed.
Removed the biome.json exclusion entry. Biome trimmed each file's imports to
actual usage. All 186 tests pass; lint clean.

Remaining: cpd surfaces 49 clones (0.3%) the monolith previously hid (jscpd
skips very large files). Next commit extracts shared helpers to dedup.
