+++ biome.json
@@ -156,7 +156,6 @@
         "src/shared/db/migrations/schema.ts",
         "src/shared/db/questions.ts",
         "src/ui/templates/public/reservations.tsx",
-        "test/lib/server-attendees.test.ts",
         "test/lib/server-listing-parents.test.ts",
         "test/lib/server-parents-gate.test.ts",
         "test/lib/server-payments.test.ts",
