+++ test/lib/server-attendees/delete.test.ts
@@ -1,0 +1,336 @@
+// jscpd:ignore-start
+import { expect } from "@std/expect";
+import { describe, it as test } from "@std/testing/bdd";
+import { handleRequest } from "#routes";
+import { getListingWithCount } from "#shared/db/listings.ts";
+import {
+  adminAttendeeAction,
+  adminFormPost,
+  adminGet,
+  adminListingPage,
+  createPaidTestAttendee,
+  createTestAttendee,
+  createTestListing,
+  describeWithEnv,
+  expectFlash,
+  expectFlashRedirect,
+  expectHtmlResponse,
+  mockFormRequest,
+  setupListingAndLogin,
+  testCookie,
+  testCsrfToken,
+  testRequiresAuth,
+} from "#test-utils";
+
+// jscpd:ignore-end
+describeWithEnv("server (admin attendees) > delete", { db: true }, () => {
+  const deleteAction = adminAttendeeAction("delete");
+
+  describe("GET /admin/listing/:listingId/attendee/:attendeeId/delete", () => {
+    testRequiresAuth("/admin/attendees/1/delete", {
+      setup: async () => {
+        const listing = await createTestListing({
+          maxAttendees: 100,
+          thankYouUrl: "https://example.com",
+        });
+        await createTestAttendee(
+          listing.id,
+          listing.slug,
+          "John Doe",
+          "john@example.com",
+        );
+      },
+    });
+
+    test("returns 404 for non-existent listing", async () => {
+      const response = await adminGet("/admin/attendees/1/delete");
+      expect(response.status).toBe(404);
+    });
+
+    test("returns 404 for non-existent attendee", async () => {
+      await createTestListing({
+        maxAttendees: 100,
+        thankYouUrl: "https://example.com",
+      });
+
+      const response = await adminGet("/admin/attendees/999/delete");
+      expect(response.status).toBe(404);
+    });
+
+    test("returns 404 for an orphan attendee with no home listing", async () => {
+      // The attendee-scoped action loads the attendee's home listing; an
+      // attendee whose bookings are all gone has none, so the page 404s.
+      const listing = await createTestListing({
+        maxAttendees: 100,
+        name: "Listing 1",
+        thankYouUrl: "https://example.com",
+      });
+      const attendee = await createTestAttendee(
+        listing.id,
+        listing.slug,
+        "John Doe",
+        "john@example.com",
+      );
+      const { getDb } = await import("#shared/db/client.ts");
+      await getDb().execute(
+        "DELETE FROM listing_attendees WHERE attendee_id = ?",
+        [attendee.id],
+      );
+
+      const response = await adminGet(`/admin/attendees/${attendee.id}/delete`);
+      expect(response.status).toBe(404);
+    });
+
+    test("shows delete confirmation page when authenticated", async () => {
+      const { response } = await adminListingPage(
+        (ctx) => `/admin/attendees/${ctx.attendee.id}/delete`,
+      )();
+      await expectHtmlResponse(
+        response,
+        200,
+        "Delete Attendee",
+        "John Doe",
+        "type their name",
+        'checked name="release_bookings" type="checkbox" value="1"',
+        "Release their bookings into the pool",
+      );
+    });
+
+    test("includes return_url as hidden field when provided", async () => {
+      const { response } = await adminListingPage(
+        (ctx) =>
+          `/admin/attendees/${ctx.attendee.id}/delete?return_url=${encodeURIComponent(
+            "/admin/calendar#attendees",
+          )}`,
+      )();
+      await expectHtmlResponse(
+        response,
+        200,
+        'name="return_url"',
+        "/admin/calendar#attendees",
+      );
+    });
+  });
+
+  describe("POST /admin/listing/:listingId/attendee/:attendeeId/delete", () => {
+    testRequiresAuth("/admin/attendees/1/delete", {
+      body: {
+        confirm_identifier: "John Doe",
+      },
+      method: "POST",
+      setup: async () => {
+        const listing = await createTestListing({
+          maxAttendees: 100,
+          thankYouUrl: "https://example.com",
+        });
+        await createTestAttendee(
+          listing.id,
+          listing.slug,
+          "John Doe",
+          "john@example.com",
+        );
+      },
+    });
+
+    test("returns 404 for non-existent listing", async () => {
+      const { response } = await adminFormPost("/admin/attendees/1/delete", {
+        confirm_identifier: "John Doe",
+      });
+      expect(response.status).toBe(404);
+    });
+
+    test("returns 404 for non-existent attendee", async () => {
+      await createTestListing({
+        maxAttendees: 100,
+        thankYouUrl: "https://example.com",
+      });
+
+      const { response } = await adminFormPost("/admin/attendees/999/delete", {
+        confirm_identifier: "John Doe",
+      });
+      expect(response.status).toBe(404);
+    });
+
+    test("rejects invalid CSRF token", async () => {
+      const { response } = await deleteAction({
+        confirm_identifier: "John Doe",
+        csrf_token: "invalid-token",
+      })();
+      await expectHtmlResponse(response, 403, "Invalid CSRF token");
+    });
+
+    test("rejects mismatched attendee name", async () => {
+      const { response } = await deleteAction({
+        confirm_identifier: "Wrong Name",
+      })();
+      expect(response.status).toBe(302);
+      expectFlash(response, expect.stringContaining("does not match"), false);
+    });
+
+    test("preserves return_url on mismatched attendee name", async () => {
+      const { response } = await deleteAction({
+        confirm_identifier: "Wrong Name",
+        return_url: "/admin/calendar#attendees",
+      })();
+      expect(response.status).toBe(302);
+      expectFlash(response, expect.stringContaining("does not match"), false);
+    });
+
+    test("deletes attendee with matching name (case insensitive)", async () => {
+      const { response, listing, attendee } = await deleteAction({
+        confirm_identifier: "john doe",
+        release_bookings: "1",
+      })();
+      await expectFlashRedirect(
+        "/admin/attendees",
+        "Attendee deleted",
+      )(response);
+
+      // Verify attendee was deleted
+      const { getAttendeeRaw } = await import("#shared/db/attendees.ts");
+      const deleted = await getAttendeeRaw(attendee.id);
+      expect(deleted).toBeNull();
+      expect(await getListingWithCount(listing.id)).toMatchObject({
+        attendee_count: 0,
+      });
+    });
+
+    test("deletes attendee with whitespace-trimmed name", async () => {
+      const { response } = await deleteAction({
+        confirm_identifier: "  John Doe  ",
+      })();
+      await expectFlashRedirect(
+        "/admin/attendees",
+        "Attendee deleted",
+      )(response);
+    });
+
+    test("can delete attendee without releasing bookings", async () => {
+      const listing = await createTestListing({ maxAttendees: 50 });
+      const attendee = await createPaidTestAttendee(
+        listing.id,
+        "Keep Pool",
+        "keep-pool@example.com",
+        "pay_keep_pool",
+        1200,
+        3,
+      );
+
+      const { response } = await adminFormPost(
+        `/admin/attendees/${attendee.id}/delete`,
+        { confirm_identifier: "Keep Pool" },
+      );
+
+      await expectFlashRedirect(
+        "/admin/attendees",
+        "Attendee deleted",
+      )(response);
+      const updated = await getListingWithCount(listing.id);
+      expect(updated).toMatchObject({
+        attendee_count: 3,
+        income: 1200,
+        tickets_count: 1,
+      });
+    });
+  });
+
+  describe("DELETE /admin/listing/:listingId/attendee/:attendeeId/delete", () => {
+    test("deletes attendee with DELETE method", async () => {
+      const listing = await createTestListing({
+        maxAttendees: 100,
+        thankYouUrl: "https://example.com",
+      });
+      const attendee = await createTestAttendee(
+        listing.id,
+        listing.slug,
+        "John Doe",
+        "john@example.com",
+      );
+
+      const formBody = new URLSearchParams({
+        confirm_identifier: "John Doe",
+        csrf_token: await testCsrfToken(),
+      }).toString();
+
+      const response = await handleRequest(
+        new Request(`http://localhost/admin/attendees/${attendee.id}/delete`, {
+          body: formBody,
+          headers: {
+            "content-type": "application/x-www-form-urlencoded",
+            cookie: await testCookie(),
+            host: "localhost",
+          },
+          method: "DELETE",
+        }),
+      );
+      await expectFlashRedirect(
+        "/admin/attendees",
+        "Attendee deleted",
+      )(response);
+
+      // Verify attendee was deleted
+      const { getAttendeeRaw } = await import("#shared/db/attendees.ts");
+      const deletedAttendee = await getAttendeeRaw(1);
+      expect(deletedAttendee).toBeNull();
+    });
+  });
+
+  describe("POST /admin/listing/:listingId/attendee/:attendeeId/delete (confirm_identifier edge case)", () => {
+    test("handles missing confirm_identifier field (falls back to empty string)", async () => {
+      // Submit without confirm_identifier field at all
+      const { response } = await deleteAction({})();
+      // Empty string won't match "John Doe", so it redirects with error
+      expect(response.status).toBe(302);
+      expectFlash(response, expect.stringContaining("does not match"), false);
+    });
+  });
+
+  describe("routes/admin/attendees.ts (parseAttendeeIds)", () => {
+    test("returns 404 for non-existent attendee on delete page", async () => {
+      const { listing, cookie } = await setupListingAndLogin({
+        maxAttendees: 50,
+        name: "Att Del 404",
+      });
+
+      const response = await handleRequest(
+        new Request(
+          `http://localhost/admin/listing/${listing.id}/attendee/99999/delete`,
+          {
+            headers: {
+              cookie,
+              host: "localhost",
+            },
+          },
+        ),
+      );
+      expect(response.status).toBe(404);
+    });
+  });
+
+  describe("routes/admin/attendees.ts (parseAttendeeIds)", () => {
+    test("exercises parseAttendeeIds via POST route with valid params", async () => {
+      const { listing, cookie, csrfToken } = await setupListingAndLogin({
+        maxAttendees: 50,
+        name: "Parse Ids Test",
+      });
+      const attendee = await createTestAttendee(
+        listing.id,
+        listing.slug,
+        "Test User",
+        "test@example.com",
+      );
+
+      // POST route exercises attendeeDeleteHandler which calls parseAttendeeIds.
+      // The custom handler requires confirm_identifier to match the attendee name.
+      const response = await handleRequest(
+        mockFormRequest(
+          `/admin/attendees/${attendee.id}/delete`,
+          { confirm_identifier: "Test User", csrf_token: csrfToken },
+          cookie,
+        ),
+      );
+      // Should redirect after successful delete
+      expect(response.status).toBe(302);
+    });
+  });
+});
